Use io.CopyN to utilize zero copy technique.

+++ recording/config.go
@@ -22,7 +22,7 @@ type TransportConfig struct {
 
 type DownloadConfig struct {
 	SaveDirectory                    string `mapstructure:"save_directory"`
-	DiskWriteBufferBytes             int    `mapstructure:"disk_write_buffer_bytes"`
+	DiskWriteBufferBytes             int64  `mapstructure:"disk_write_buffer_bytes"`
 	UseSpecialExtNameBeforeFinishing bool   `mapstructure:"use_special_ext_name_when_downloading"`
 }
